pyproject.toml
[project]
name = "autobiota"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.1",
    "xgboost>=3.0.2",
]
